upload api improved

+++ routes/product.js
@@ -44,7 +44,7 @@ router.post(
     },
   ]),
   async (req, res) => {
-    const { uid, title, size, desc } = req.body;
+    const { uid, title, desc } = req.body;
     const price = JSON.parse(req.body.price);
     const filters = JSON.parse(req.body.filters);
     const tags = JSON.parse(req.body.tags);
@@ -149,6 +149,178 @@ router.post(
       });
   }
 );
+router.post(
+  "/update",
+  fileUpload.fields([
+    {
+      name: "thumbnail",
+      maxCount: 1,
+    },
+    {
+      name: "image1",
+      maxCount: 1,
+    },
+    {
+      name: "image2",
+      maxCount: 1,
+    },
+    {
+      name: "image3",
+      maxCount: 1,
+    },
+  ]),
+  async (req, res) => {
+    const { uid, pid, title, desc } = req.body;
+    const price = JSON.parse(req.body.price);
+    
+        let thumbnail, image1, image2, image3;
+    
+        if (req.files.thumbnail) thumbnail = req.files.thumbnail[0].path;
+        if (req.files.image1) image1 = req.files.image1[0].path;
+        if (req.files.image2) image2 = req.files.image2[0].path;
+        if (req.files.image3) image3 = req.files.image3[0].path;
+    
+    if (!uid) {
+      if(thumbnail)
+      await unlinkFileAsync(thumbnail);
+      if(image1)
+      await unlinkFileAsync(image1);
+      if(image2)
+      await unlinkFileAsync(image2);
+      if(image3)
+      await unlinkFileAsync(image3);
+    
+      res.status(422).json({
+        status: false,
+        message: "User's Id not provided",
+      });
+      return;
+    }
+    if (!pid) {
+      if(thumbnail)
+      await unlinkFileAsync(thumbnail);
+      if(image1)
+      await unlinkFileAsync(image1);
+      if(image2)
+      await unlinkFileAsync(image2);
+      if(image3)
+      await unlinkFileAsync(image3);
+      res.status(422).json({
+        status: false,
+        message: "Product's Id not provided",
+      });
+      return;
+    }
+    let result;
+    try {
+      result = await User.findOne({ _id: uid }, "-password");
+    } catch {async() => {
+        if(thumbnail)
+        await unlinkFileAsync(thumbnail);
+        if(image1)
+        await unlinkFileAsync(image1);
+        if(image2)
+        await unlinkFileAsync(image2);
+        if(image3)
+        await unlinkFileAsync(image3);
+        res.status(422).json({
+          status: false,
+          message: "Invalid id",
+        });
+      }}
+
+    if (!result) {
+      if(thumbnail)
+      await unlinkFileAsync(thumbnail);
+      if(image1)
+      await unlinkFileAsync(image1);
+      if(image2)
+      await unlinkFileAsync(image2);
+      if(image3)
+      await unlinkFileAsync(image3);
+      res.status(422).json({
+        status: false,
+        message: "User not found in our database",
+      });
+      return;
+    }
+
+    if (result.role != "admin") {
+      if(thumbnail)
+      await unlinkFileAsync(thumbnail);
+      if(image1)
+      await unlinkFileAsync(image1);
+      if(image2)
+      await unlinkFileAsync(image2);
+      if(image3)
+      await unlinkFileAsync(image3);
+      res.status(422).json({
+        status: false,
+        message: "User do not have permission",
+      });
+      return;
+    }
+
+    const product = await Product.findOne({ _id: pid });
+
+    if (title) {
+      product.title = title;
+    }
+    if (price) {
+      product.price = price;
+    }
+    if (desc) {
+      product.desc = desc;
+    }
+    // if (tags) {
+    //   product.tags = tags;
+    // }
+    // if (filters) {
+    //   product.filters = {
+    //     type: filters.type,
+    //     season: filters.season,
+    //     for: filters.for,
+    //     brand: filters.brand,
+    //   };
+    // }
+    if (thumbnail) {
+      product.image = thumbnail;
+      product.images.splice(0, 1, thumbnail);
+    }
+    if (image1) {
+      product.images.splice(1, 1, image1);
+    }
+    if (image2) {
+      product.images.splice(2, 1, image2);
+    }
+    if (image3) {
+      product.images.splice(3, 1, image3);
+    }
+   ;
+    product.save()
+      .then(() => {
+        res.status(201).json({
+          status: true,
+          message: "Product Updated",
+        });
+      })
+      .catch(async (err) => {
+        if(thumbnail)
+        await unlinkFileAsync(thumbnail);
+        if(image1)
+        await unlinkFileAsync(image1);
+        if(image2)
+        await unlinkFileAsync(image2);
+        if(image3)
+        await unlinkFileAsync(image3);
+        res.status(502).json({
+          status: false,
+          message: `Error updating Product`,
+          error: err,
+        });
+      });
+  }
+);
 
 router.post("/filter-search", async (req, res) => {
   const filters = req.body.filters;
@@ -207,8 +379,17 @@ router.post("/filter-search", async (req, res) => {
   return;
 });
 
-router.delete("/:id", (req, res) => {
-  const id = req.params.id;
+router.post("/delete", (req, res) => {
+  const id = req.body.id;
+
+  if (!id) {
+    res.status(422).json({
+      status: false,
+      message: "Product's Id not provided",
+    });
+    return;
+  }
+
   Product.deleteOne({ _id: id })
     .then(() => {
       res.status(200).json({
